finish reOrganize Card Component
