+++ src/components/CardChooseDeleteContainer.jsx
@@ -1,0 +1,11 @@
+import { CardDeleteBtn } from "./CardDeleteBtn"
+import { CardChooseBtn } from "./CardChooseBtn"
+
+export const CardChooseDelete = ({id}) => {
+    return (
+        <div className='chooseDeleteContainer'>
+            <CardDeleteBtn id={id}/>
+            <CardChooseBtn id={id}/>
+        </div>
+    )
+}
