+++ src/utils/cardUtils.js
@@ -1,0 +1,7 @@
+export const isLocationExist = (currCardInfo) => {
+    const LocationValues = Object.values(currCardInfo.location)
+    if (LocationValues.length) {
+        return true
+    }
+    return false
+}
